+++ nomos-services/data-availability/network/src/membership/mod.rs
@@ -1,4 +1,4 @@
-pub mod adapter;
+pub mod adapters;
 pub mod handler;
 
 use std::collections::{HashMap, HashSet};
