+++ nomos-services/membership/src/backends/mock.rs
@@ -11,7 +11,7 @@ use nomos_sdp_core::{
 use super::{MembershipBackend, MembershipBackendError, Settings};
 use crate::MembershipProviders;
 
-#[derive(Clone)]
+#[derive(Clone, Debug)]
 pub struct MockMembershipBackendSettings {
     settings_per_service: HashMap<ServiceType, Settings>,
     initial_membership: HashMap<BlockNumber, MockMembershipEntry>,
