+++ nomos-services/membership/src/backends/mod.rs
@@ -14,7 +14,7 @@ pub struct MembershipBackendSettings {
     pub settings_per_service: HashMap<ServiceType, Settings>,
 }
 
-#[derive(Clone)]
+#[derive(Clone, Debug)]
 pub struct Settings {
     historical_block_delta: u64,
 }
