+++ nomos-services/sdp/src/adapters/services/mod.rs
@@ -1,5 +1,1 @@
 pub mod services_repository;
-
-pub trait SdpServicesAdapter {
-    fn new() -> Self;
-}
