+++ nomos-services/sdp/src/adapters/services/services_repository.rs
@@ -4,7 +4,7 @@ use nomos_sdp_core::ledger;
 #[derive(Debug, Clone)]
 pub struct LedgerServicesAdapter;
 
-pub trait SdpServicesAdapter {
+pub trait SdpServicesAdapter: ledger::ServicesRepository {
     fn new() -> Self;
 }
 impl SdpServicesAdapter for LedgerServicesAdapter {
