+++ nomos-services/sdp/src/backends/ledger.rs
@@ -1,16 +1,15 @@
 use nomos_core::block::BlockNumber;
-use nomos_sdp_core::{
-    ledger::{SdpLedger, ServicesRepository},
-    SdpMessage,
-};
+use nomos_sdp_core::{ledger::SdpLedger, SdpMessage};
 
 use super::{SdpBackend, SdpBackendError};
-use crate::adapters::{declaration::SdpDeclarationAdapter, services::SdpServicesAdapter};
+use crate::adapters::{
+    declaration::repository::SdpDeclarationAdapter,
+    services::services_repository::SdpServicesAdapter,
+};
 
 #[async_trait::async_trait]
 impl<Declarations, Services, Metadata> SdpBackend for SdpLedger<Declarations, Services, Metadata>
 where
-    Services: ServicesRepository + Send + Sync + Clone + 'static,
     Metadata: Send + Sync + 'static,
     Declarations: SdpDeclarationAdapter + Send + Sync,
     Services: SdpServicesAdapter + Send + Sync,
